Prevent CORS errors when using internal styles
Due to a new-ish CORS rule in many browsers (including Chromium, but I've also replicated this in Safari and Firefox), the cssRules property can no longer be read from external stylesheets, and attempts to do so throw a security error: "Failed to read the 'cssRules' property from 'CSSStyleSheet': Cannot access rules". This can occur with users who have certain browser extensions enabled, or when using external webfonts, etc.

We can prevent this issue by filtering out any stylesheets that are not on the same origin when reading their rules. Unfortunately this may in some cases lead to less accurate styles, but that's an inherent downside of this approach, and developers should be aware of the tradeoff.